Allow the code generator to be invoked separately
Passing input that ends in `.json` instead of `.salt` will cause the front-end
to be skipped, and the input interpreted as slugs json output. When using this
input form, you must also specify the `--length` argument, which instructs the
code generator to treat that many parameters in the controller state as inputs.